Activate magithub-minor-mode more conservatively
Replace "magithub-try-enabling-minor-mode-everywhere" with
"magithub-try-enabling-minor-mode-for-repo", which only tries to
enable magithub-minor-mode for buffers corresponding to files in the
current repo.

This solves the issue with magithub hanging indefinitely when you have
a remote file open on a currently-inaccessible server. In such cases,
magithub would try to enable its minro mode in the remote file's
buffer, which would cause TRAMP to try to reconnect to the server,
which would hang because the server was inaccessible.